refactor(spf): move DeriveStartMediaTime seam to a shared primitive
establishStartMediaTime and relocation-pipelines each declared an identical
DeriveStartMediaTime type. Extract it (plus DeriveStartMediaTimeContext) to a
new playback/primitives/derive-start-media-time — the lowest layer both reach
(behaviors -> primitives is allowed, the reverse is not). It's a playback-domain
contract (it names selected track ids), so it belongs at the playback layer, not
in the framework-agnostic media/. The behavior re-exports it, so engine and
public exports are unchanged.
